backend/requirements.txt
django 
djangorestframework 
celery
redis
pandas
gunicorn 
psycopg2-binary
djangorestframework-simplejwt
Pillow
